Fix pointer cast warnings, compliant C99 _Bool (4f9a1c9 upstream)
agbcc now can use _Bool. It works like _Bool, not an enum.

+++ gcc/rtl.h
@@ -1117,7 +1117,7 @@ extern rtx gen_rtx_MEM(enum machine_mode, rtx);
 
 /* We need the cast here to ensure that we get the same result both with
    and without prototypes.  */
-#define GEN_INT(N)  gen_rtx_CONST_INT (VOIDmode, (int32_t) (N))
+#define GEN_INT(N)  gen_rtx_CONST_INT (VOIDmode, (intptr_t) (N))
 
 #define arg_pointer_rtx		(&global_rtl.arg_pointer_val)
